Added additional tests for undefined.u1
@thejoshwolfe pointed out that since u1 is currently stored in at least
a u8 it is incorrect to assume a u1 has only two values.

I changed the test to have a else => assert(@truncate(u1, val) == 1),
and it now works.

I added some additional tests for undefined.u1 and they all work in
all the --release-xxxx builds but fails in a debug build, this seems odd.