Never read undefined data.
One of the reasons I was reading undefined value was I wanted to test
initializing val to a random value from the uninitialized data, which
was an incorrect assumption that Josh Wolfe enlightened me.

So I created rand() and used it with @truncate to initialize,
`val = @truncate(u1, rand());`. That worked fine for u1 but fails for
u0. For now I use `val = @truncate(u0, 123);` which isn't what I want,
but good enough for now.